% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{branch_k2}
\alias{branch_k2}
\title{Calculate the length of a branch that ends with coalescence of two samples (i.e., coal rate = 1)}
\usage{
branch_k2(t0 = 0, tm = Inf)
}
\arguments{
\item{t0}{time to beginning of ancestral lineage (i.e., a constant branch length to add) in coalescent units}

\item{tm}{maximum time to coalescence (total time of ancestral lineage) in coalescent units}
}
\value{
Numeric, a branch length in coalescent units
}
\description{
Calculate the length of a branch that ends with coalescence of two samples (i.e., coal rate = 1)
}
