% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{branch_length_correlation}
\alias{branch_length_correlation}
\title{Correlation between number of substitutions and coalescent length units}
\usage{
branch_length_correlation(mpest_branches, concate_branches)
}
\arguments{
\item{mpest_branches}{vector of coalescent lengths}

\item{concate_branches}{vector of substitutions}
}
\value{
a vector with slope and intercept
}
\description{
Basic correlation, to be used in the generation of approximate
branch lengths in coalescent units (not estimated by coalescent tree methods)

This is an implementation of Meng Wu's idea to correct MP-EST branches.
}
