% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_functions.R
\name{calc_hrf}
\alias{calc_hrf}
\title{Calculate the Hemiplasy Risk Factor in a three-species clade.}
\usage{
calc_hrf(this_branch, descendants, ancestor, sibling, mutation_rate,
  model = "standard", pepo = F)
}
\arguments{
\item{this_branch}{Numeric. Branch length of focal internal branch.}

\item{descendants}{Numeric vector of size 2. Lengths of branches descending from focal branch.}

\item{ancestor}{Numeric. Length of ancestral branch (the ABC ancestor)}

\item{sibling}{Numeric. Length of sibling branch (leading to tip node A).}

\item{mutation_rate}{Numeric (forward, reverse population-wide mutation rate)}

\item{model}{String. One of the three models implemented: "standard" (default), "minimal", or "strict".}
}
\value{
Numeric, a ratio
}
\description{
HRF is the expected fraction of incongruency due to hemiplasy (vs. homoplasy). It assumes:
no uncertainty on species tree (A,(B,C)), one sample per species, branch lengths in coalescent units.

Calls \code{pr_homoplasy()} and \code{pr_hemiplasy()} for the given values. The total probability of
each scenario is the sum of probabilities for two incongruencies:
when A and C share the derived state, and when B and C share the derived state.

If 'mutation_rate' is a numeric of length 1, forward (0->1) and reverse (1->0) mutation rates are the same.
If 'mutation_rate' is of length 2, the first element is the forward- and the second is the reverse mutation rate.

See documentation for tree_hrf() for description of the three models implemented.
}
\seealso{
\code{\link{pr_homoplasy}}, \code{\link{pr_hemiplasy}}, \code{\link{tree_hrf}}
}
