% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{convert_from_cf}
\alias{convert_from_cf}
\title{Use concordance factors as approximate branch lengths}
\usage{
convert_from_cf(tree, concordance_factors)
}
\arguments{
\item{tree}{a phylogeny of class 'phylo'}

\item{concordance_factors}{string or numeric vector of length equal to number of branches in tree}
}
\value{
numeric vector of estimated branches
}
\description{
Under the assumption that all gene tree discordance is due to
incomplete lineage sorting, this function estimates length for internal branches of a tree
by taking the proportion of discordant trees (i.e. 1 minus the concordance factor)
and solving for branch length t = -ln(3⁄2* (1-CF)).
It will return 0 for CF < 1/3, and
Infinite for CF == 1 (which isn't realistic but essentially eliminates the possibility of hemiplasy)
}
