% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{fix_mpest_tips}
\alias{fix_mpest_tips}
\title{Change values of branch lengths at tips}
\usage{
fix_mpest_tips(branches, is_tip, newval)
}
\arguments{
\item{branches}{numeric vector}

\item{is_tip}{logical vector}

\item{newval}{numeric or class 'phylo'}
}
\description{
Because estimating branch length at tips (terminal lineages) is
  impossible for coalescent methods, some implementations (like MP-EST)
  return tips of arbitrary length. These lengths don't have biological
  meaning but can affect the value of HRF. This function changes the values
  of the tips for another, equally arbitrary length. 'newval' can take a
  single value that is applied to all tips, a vector of equal length to the
  number of tips, or a phylogeny (object class 'phylo') with branch lengths
  in substitution rate (will be used to extrapolate coalescent branch length)
}
