% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{homoplasy_mutations}
\alias{homoplasy_mutations}
\title{Get probabilities of mutation along all branches}
\usage{
homoplasy_mutations(this, desc_a, desc_b, sib, anc, mu_rate, coal = T,
  mode)
}
\arguments{
\item{this}{Numeric. Branch length of focal internal branch.}

\item{desc_a}{Lengths of branches descending from focal branch.}

\item{desc_b}{Lengths of branches descending from focal branch.}

\item{sib}{Length of sibling branch (leading to tip node C).}

\item{anc}{Length of ancestral branch (the ABC ancestor)}

\item{mu_rate}{Numeric of length 2. Population-wide mutation rate (forward, reverse)}

\item{coal}{logical. Did coalescence in the BC ancestor happen?}

\item{mode}{String. One of the three models implemented: "standard" (default), "minimal", or "strict".}
}
\value{
List of size 5
}
\description{
Given a species tree, calculate the probability of mutation on each of the branches of the gene tree.
}
