% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepo.R
\docType{package}
\name{pepo}
\alias{pepo}
\alias{pepo-package}
\title{P(e)P(o): Compute the relative importance of hemiplasy and homoplasy in a phylogeny.}
\description{
P(e)P(o): Compute the relative importance of hemiplasy and homoplasy in a phylogeny.
}
\section{Tree functions}{

The tree functions ...
}

\section{Basic functions}{

The basic functions ...
}

