% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{pr_double_hit}
\alias{pr_double_hit}
\title{Calculate probability of two mutations on three-species tree}
\usage{
pr_double_hit(mu_list, where)
}
\arguments{
\item{mu_list}{List with "anc", "a", "b", "c", and "ab" entries, each of which is a mutation probability.}

\item{where}{String. "ac" or "banc".}
}
\value{
Numeric. Probability
}
\description{
One of two types of double-mutation events will be returned, depending on the value of \code{where}.
Mutation and back-mutation are assumed to have the same rate.

"ac"--> product of the following probabilities: mutation on the [a] and [c] branches, no mutations on [b], ancestral, or internal branches.
"banc" --> product of the following probabilities: mutation on the [abc] ancestor and on the [b] branch, no mutation on [a],[c] or internal branches.
}
\seealso{
\code{\link{homoplasy_mutations}}
}
