% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_functions.R
\name{pr_hemiplasy}
\alias{pr_hemiplasy}
\title{Probability of incongruence due to hemiplasy in a three-species clade.}
\usage{
pr_hemiplasy(this, desc_a, desc_b, sib, anc, mu_rate, mode = "standard")
}
\arguments{
\item{this}{Numeric. Branch length of focal internal branch.}

\item{desc_a, desc_b}{Lengths of branches descending from focal branch.}

\item{sib}{Length of sibling branch (leading to tip node C).}

\item{anc}{Length of ancestral branch (the ABC ancestor)}

\item{mu_rate}{Numeric of length 2. Population-wide mutation rate (forward, reverse)}

\item{mode}{String. One of the three models implemented: "standard" (default), "minimal", or "strict".}
}
\value{
Numeric, a probability
}
\description{
It assumes:
no uncertainty on species tree (A,(B,C)), one sample per species, observed states of
binary trait were a=1, b=0, and c=1, branch lengths in coalescent units.

Roughly, it computes the product of the probabilities of events necessary for the observed
incongruence. Specifically, the probability of samples a and c coalescing (which includes incomplete
lineage sorting in the internal BC branch), and one mutation on the ac branch (which implies no mutations
on other branches of the tree).

See documentation of tree_hrf() for description of the three models implemented.
}
\seealso{
\code{\link{pr_homoplasy}}, \code{\link{single_hrf}}, \code{\link{tree_hrf}}
}
