% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_functions.R
\name{pr_homoplasy}
\alias{pr_homoplasy}
\title{Probability of incongruence due to homoplasy in a three-species clade.}
\usage{
pr_homoplasy(this, desc_a, desc_b, sib, anc, mu_rate, mode = "standard")
}
\arguments{
\item{this}{Numeric. Branch length of focal internal branch.}

\item{desc_a}{Lengths of branches descending from focal branch.}

\item{desc_b}{Lengths of branches descending from focal branch.}

\item{sib}{Length of sibling branch (leading to tip node C).}

\item{anc}{Length of ancestral branch (the ABC ancestor)}

\item{mu_rate}{Numeric of length 2. Population-wide mutation rate (forward, reverse)}

\item{mode}{String. One of the three models implemented: "standard" (default), "minimal", or "strict".}
}
\value{
Numeric, a probability
}
\description{
It assumes:
no uncertainty on species tree (A,(B,C)), one sample per species, observed states of
binary trait were a=1, b=0, and c=1, branch lengths in coalescent units.

Roughly, it computes the product of the probabilities of events necessary for the observed
incongruence. Specifically, mutations along the A and C branches (implying no mutations
in other branches) or --depending on \code{mode}- mutations along the ABC ancestor and B.

See documentation of tree_hrf() for description of the three models implemented.
}
\seealso{
\code{\link{pr_hemiplasy}}, \code{\link{calc_hrf}}, \code{\link{tree_hrf}}
}
