% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{pr_mutation}
\alias{pr_mutation}
\title{Calculate the probability of at least one mutation on a branch}
\usage{
pr_mutation(mu, t)
}
\arguments{
\item{mu}{mutation rate in coalescent units, 2N*per_generation_mutation_rate}

\item{t}{branch length, in coalescent units}
}
\value{
Numeric, the probability of at least one mutation.
}
\description{
Calculate the probability of at least one mutation on a branch
}
\examples{
pr_mutation(0.001, 1)
}
