% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_functions.R
\name{single_hrf}
\alias{single_hrf}
\title{Ratio of hemiplasy to homoplasy in a three-species clade.}
\usage{
single_hrf(this_branch, descendants, ancestor, sibling, mutation_rate,
  model = "standard")
}
\arguments{
\item{this_branch}{Branch length of focal internal branch.}

\item{descendants}{Numeric vector of size 2. Lengths of branches descending from focal branch.}

\item{ancestor}{Length of ancestral branch (the ABC ancestor)}

\item{sibling}{Length of sibling branch (leading to tip node C).}

\item{mutation_rate}{Population-wide mutation rate}

\item{model}{One of three values: "minimal" assumes no back-mutation, so the probability of observing}
}
\value{
Numeric, a ratio
}
\description{
It assumes:
no uncertainty on species tree ((A,B),C), one sample per species, branch lengths in coalescent units.

It calls \code{pr_homoplasy()} and \code{pr_hemiplasy()} for the given values. If \code{both_sides == TRUE}
(default), it calculates probabilities for two incongruencies: when A and C share the derived state, and when
B and C share the derived state. Otherwise, it only calculates the first case (it assumes that the
first value in the \code{descendants} vector is A.
}
