% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{to_phylo4d}
\alias{to_phylo4d}
\title{Convert to 'phylo4' class}
\usage{
to_phylo4d(tree, df)
}
\arguments{
\item{tree}{an object of class "phylo"}

\item{df}{accesory data to be added}
}
\value{
an object of class phylo4
}
\description{
Convert to 'phylo4' class
}
