% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{tree_hrf}
\alias{tree_hrf}
\title{Calculate HRF for all branches on a tree}
\usage{
tree_hrf(edges, mutation = 0.01, mode = "standard")
}
\arguments{
\item{mode}{}
}
\value{
a dplyr::data_frame
}
\description{
Calculate HRF for all branches on a tree
}
