% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{branch_k2}
\alias{branch_k2}
\title{Calculate the length of a branch that ends with coalescence of two samples (i.e., coal rate = 1)}
\usage{
branch_k2(t0 = 0, tm = Inf)
}
\arguments{
\item{t0}{Numeric. Time at start of coalescent process (i.e., constant length to be added to terminal branches; in coal units)}

\item{tm}{Numberic. Maximum time to coalescence (ABC ancestor branch length)}
}
\value{
Numeric, a branch length in coalescent units
}
\description{
Calculate the length of a branch that ends with coalescence of two samples (i.e., coal rate = 1)
}
