% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_functions.R
\name{branch_k3}
\alias{branch_k3}
\title{Calculate the length of a branch that ends in a population where three samples coalesce}
\usage{
branch_k3(t0 = 0, tm = Inf, branch = "avg")
}
\arguments{
\item{t0}{time to ancestor, or time to speciation event (i.e., a constant branch length, in coalescent units)}

\item{tm}{maximum time to coalescence (ancestral branch length)}

\item{branch}{specifies if the time to the first ('short'), second ('long'), internal ('inter'), ancestral ('anc'), mean ('avg') coalescent event.}
}
\value{
Numeric, the probability of at least one mutation.
}
\description{
Calculate the length of a branch that ends in a population where three samples coalesce
}
\examples{
branch_k3(1, 10, 'short')
}
