% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{prep_branch_lengths}
\alias{prep_branch_lengths}
\title{Prepare branches for HRF calculation}
\usage{
prep_branch_lengths(tree)
}
\arguments{
\item{tree}{}
}
\value{
a dplyr::data_frame
}
\description{
Prepare branches for HRF calculation
}
