% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{prep_branch_lengths}
\alias{prep_branch_lengths}
\title{Prepare branches for HRF calculation}
\usage{
prep_branch_lengths(tree, tip_replace = NULL)
}
\arguments{
\item{tree}{a phylogeny of class 'phylo'}
}
\value{
a tibble
}
\description{
Takes a 'phylo' object and extracts the relevant branches for each HRF calculation.
}
