% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_functions.R
\name{single_hrf}
\alias{single_hrf}
\title{Ratio of hemiplasy to homoplasy in a three-species clade.}
\usage{
single_hrf(this_branch, descendants, ancestor, sibling, mutation_rate,
  model = "standard", pepo = F)
}
\arguments{
\item{this_branch}{Numeric. Branch length of focal internal branch.}

\item{descendants}{Numeric vector of size 2. Lengths of branches descending from focal branch.}

\item{ancestor}{Numeric. Length of ancestral branch (the ABC ancestor)}

\item{sibling}{Numeric. Length of sibling branch (leading to tip node C).}

\item{mutation_rate}{Numeric. Population-wide mutation rate}

\item{model}{String. One of the three models implemented: "standard" (default), "minimal", or "strict".}
}
\value{
Numeric, a ratio
}
\description{
It assumes:
no uncertainty on species tree ((A,B),C), one sample per species, branch lengths in coalescent units.

Calls \code{pr_homoplasy()} and \code{pr_hemiplasy()} for the given values. If \code{both_sides == TRUE}
(default), the total probability of each scenario is the sum of probabilities for two incongruencies:
when A and C share the derived state, and when B and C share the derived state.
Otherwise, it only calculates the first case (assuming the first value in the
\code{descendants} vector as A).

See documentation for tree_hrf() for description of the three models implemented.
}
\seealso{
\code{\link{pr_homoplasy}}, \code{\link{pr_hemiplasy}}, \code{\link{tree_hrf}}
}
