% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{to_treedata}
\alias{to_treedata}
\title{Convert to 'treedata' structure}
\usage{
to_treedata(tree, df)
}
\arguments{
\item{tree}{an object of class "phylo"}

\item{df}{accesory data to be added}
}
\value{
an object of class treedata
}
\description{
Convert to 'treedata' structure
}
