% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_on_trees.R
\name{tree_hrf}
\alias{tree_hrf}
\title{Calculate HRF for all branches on a tree}
\usage{
tree_hrf(edges, mutation = 0.01, mode = "standard")
}
\arguments{
\item{edges}{a tbl (or data_frame)}

\item{mutation}{numeric, population-wide mutation rate}

\item{mode}{string}
}
\value{
a tibble
}
\description{
It assumes:
no uncertainty on species tree ((A,B),C), one sample per species, branch lengths in coalescent units.
}
\seealso{
\code{\link{pr_homoplasy}}, \code{\link{pr_hemiplasy}}, \code{\link{single_hrf}}
}
